# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoCRM API",
    "description": "Autotherm project lifecycle system. Money is integer minor units with an explicit currency; quantities are decimal strings. PATCH: omit a field to keep it, send null to clear it.",
    "license": {
      "name": ""
    },
    "version": "1"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/orders": {
      "get": {
        "tags": [
          "orders"
        ],
        "operationId": "orders_search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Matches number, title, partner, VIN, and plate ignoring spaces and dashes.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "description": "Stage key.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "partner_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "project_type_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "assigned_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "open",
            "in": "query",
            "description": "Only orders not in a terminal stage.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_OrderSummary"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "orders"
        ],
        "operationId": "orders_create",
        "requestBody": {
          "description": "`partner_id` and `title` are required.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "orders"
        ],
        "operationId": "orders_detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "orders"
        ],
        "operationId": "orders_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchOrder"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/stage": {
      "post": {
        "tags": [
          "orders"
        ],
        "operationId": "orders_change_stage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageChange"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/stages": {
      "get": {
        "tags": [
          "orders"
        ],
        "operationId": "orders_stage_history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stage history, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_StageEntry"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Currency": {
        "type": "string",
        "enum": [
          "HUF",
          "EUR"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "OrderBody": {
        "type": "object",
        "required": [
          "currency"
        ],
        "properties": {
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "partner_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "contact_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "project_type_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "valuation_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date",
            "description": "Defaults to today. Reports normalise the order's value at this date's MNB rate."
          },
          "vehicle_make": {
            "type": [
              "string",
              "null"
            ]
          },
          "vehicle_model": {
            "type": [
              "string",
              "null"
            ]
          },
          "vehicle_plate": {
            "type": [
              "string",
              "null"
            ]
          },
          "vehicle_vin": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "due_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "assigned_to": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemBody"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "autocrm_session"
      }
    }
  },
  "security": [
    {
      "session_cookie": []
    },
    {
      "bearer": []
    }
  ],
...
}